Add half-day leave support
- Half Day checkbox appears automatically when start and end date are the same day
- Ticking it sets days to 0.5 instead of 1
- Unchecking or changing dates resets to full days
- DB schema updated: days column changed from INTEGER to REAL

+++ backend/db.js
@@ -138,7 +138,7 @@ function createTables(db) {
       leave_type TEXT CHECK(leave_type IN ('annual','mc','emergency','unpaid','maternity','paternity')),
       start_date TEXT NOT NULL,
       end_date TEXT NOT NULL,
-      days INTEGER NOT NULL,
+      days REAL NOT NULL,
       reason TEXT,
       document_path TEXT,
       status TEXT DEFAULT 'pending' CHECK(status IN ('pending','approved','rejected')),
